// pages/on_board/types.ts
/****************************************************************************\
 *                   © ViVuDi 2025. All rights reserved.                    *
 ****************************************************************************
 *  File        : Chứa types sử dụng trong OnBoardScreen                    *
 *  Author      : Minh Nhat                                                 *
 *  Created     : 14/06/2025                                                *
 *  Updated by  :                                                           *
 *  Modified    :                                                           *
\****************************************************************************/

import { ImageSourcePropType } from "react-native";

/****************************************************************************
* Slide: Định nghĩa kiểu dữ liệu cho từng slide trong OnBoardScreen         *
****************************************************************************/
export type Slide = {
    /****************************************************************************
     * Khóa định danh duy nhất cho slide                                        *
     ****************************************************************************/
    key: string;
    /****************************************************************************
     * Ảnh hiển thị trên slide                                                  *
     ****************************************************************************/
    image: ImageSourcePropType;
    /****************************************************************************
     * Tiêu đề của slide (2 phần: bình thường và highlight)                     *
     ****************************************************************************/
    title: [string, string];
    /****************************************************************************
     * Màu sắc highlight cho tiêu đề                                            *
     ****************************************************************************/
    highlightColor: string;
    /****************************************************************************
     * Mô tả chi tiết cho slide                                                 *
     ****************************************************************************/
    description: string;
    /****************************************************************************
     * Màu nền cho button và slide dot bên dưới                                 *
     ****************************************************************************/
    buttonColor: string;
    /****************************************************************************
     * Độ dài các chấm chỉ số slide                                             *
     ****************************************************************************/
    dotLength: number[];
    /****************************************************************************
     * Vị trí đường cong bên trái                                               *
     ****************************************************************************/
    curveLeftPosition: number;
    /****************************************************************************
     * Màu sắc cho các chấm không hoạt động                                     *
     ****************************************************************************/
    inactiveColor: string;
};